refactor: extract the mediator-config crate — raw TOML schema (simplification T18, part a)
Phase 4 (config unification) begins. The mediator's TOML *schema* — the
`*ConfigRaw` serde types that mirror `mediator.toml` — moves into a new,
dependency-light `affinidi-messaging-mediator-config` crate so it can later be
shared with the `mediator-setup` wizard, which today hand-renders the same TOML
with no shared types (structural drift waiting to happen).

The config module already had a clean RAW (serde) vs RESOLVED (runtime) seam;
this extraction follows it rather than inventing one.

Moved to the crate (schema only, zero runtime deps — serde + the lean tier of
mediator-common for `DatabaseConfigRaw`):
- `ConfigRaw` + `ServerConfig` / `StreamingConfig` / `DIDResolverConfig` /
  `SecretsConfigRaw` / `StorageConfig` / `SecurityConfigRaw` / `LimitsConfigRaw` /
  `ProcessorsConfigRaw` (+ forwarding/cleanup raw structs) and serde default fns.

Stayed in the mediator (all runtime concerns):
- The resolved `Config` / `SecurityConfig` / `LimitsConfig` / `ProcessorsConfig`,
  every `ConfigRaw -> Config` conversion, secret-backend/VTA/DID-resolver
  loading, and (deferred to T18b) the env-var overrides + boot validation.
- Three conversions the orphan rule no longer allows as inherent/`TryFrom` impls
  became free fns / an extension trait, bodies verbatim:
  `did_resolver_cache_config`, `forwarding_config_from_raw`,
  `SecurityConfigRawExt::convert`.

The mediator re-exports the crate (`pub use affinidi_messaging_mediator_config::*`)
so every `crate::common::config::*` path — and external consumers like
`affinidi-messaging-test-mediator` — keep resolving unchanged.

mediator-common: `database` module un-gated from `server` so the lean crate can
embed `DatabaseConfigRaw`; only the `DatabaseConfig` `TryFrom` (returns
`MediatorError`) stays `server`-gated. Additive; server builds unchanged.

Scope note: env-var overrides + boot validation were originally slated for this
PR but both carry mediator-common's server-tier `MediatorError`; they move
together in T18b (with a crate-local config error type) rather than dragging the
server stack into the schema crate.

Verification: golden test parses the shipped conf/mediator.toml into the
relocated `ConfigRaw`; mediator lib 144, mediator-common 141, config crate 1,
test-mediator e2e all green; clippy clean on both feature sets; the config crate
builds standalone with no server-stack deps. mediator 0.15.40, mediator-common
0.15.11, mediator-config 0.1.0.
