+++ crates/messaging/affinidi-messaging-mediator/affinidi-messaging-mediator-common/src/lib.rs
@@ -24,12 +24,13 @@ pub mod types;
 
 #[cfg(feature = "server")]
 pub mod circuit_breaker;
-/// Database wiring: the lean `DatabaseConfig` serde struct is always
-/// available under `server` (the mediator's main config parses it
-/// regardless of which storage backend is selected). The
-/// `DatabaseHandler` and its redis-specific helpers are gated inside
-/// the module on `redis-backend`.
-#[cfg(feature = "server")]
+/// Database wiring. The raw serde structs in `database::config`
+/// (`DatabaseConfigRaw`, `DatabaseConfig`) are **always available** —
+/// the lean `affinidi-messaging-mediator-config` crate embeds
+/// `DatabaseConfigRaw` in its top-level `ConfigRaw` and must reach it
+/// with `default-features = false`. The `DatabaseConfig` `TryFrom`
+/// (which returns `MediatorError`) is gated on `server`, and
+/// `DatabaseHandler` + its redis helpers are gated on `redis-backend`.
 pub mod database;
 #[cfg(feature = "server")]
 pub mod errors;
